tagger/requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
numpy==1.19.0
opencv-python==4.3.0.36
Pillow==7.2.0
pydicom==2.0.0
requests==2.24.0
urllib3==1.25.9
